requirements.txt
flask
pillow
scikit-learn
numpy
requests
webcolors
opencv-python